Switch from flex layout to traditional and make nav sticky
To make the nav sticky using `position: sticky` (hopefully more robust
than what we were doing previously), we need the sidebar height to be
its content height. With the previous flex layout, it was always equal
to the height of its container, which breaks sticky positioning.

This switches to using regular old floats for layout, which gives the
sidebar its native height and allows it to be sticky positioned.